feat: refactor and port to Amazon Lambda
