+++ misc_html/openEDXsignup-Lambda.html
@@ -1,0 +1,44 @@
+<script src="https://code.jquery.com/jquery-2.2.4.min.js" crossorigin="anonymous"></script>
+<script>
+  function requestAccount() {
+    var baseUrl = window.location.href.split("/", 3).join("/");
+    $.ajax({
+      url: baseUrl + "/api/user/v1/accounts",
+      dataType: "json",
+      success: function(data) {
+        var email = data[0].email;
+        var obj = {
+          email: email,
+          url: "https://academy.aggregate.dhis2.org/tl",
+        };
+        console.log(obj);
+        $.ajax({
+          url: "https://ue7emdwas56wqbk45kp3veyaeq0azyfz.lambda-url.eu-west-1.on.aws/",
+          type: "POST",
+          data: JSON.stringify(obj),
+          contentType: "application/json; charset=utf-8",
+          dataType: "json",
+          success: function(result, status, xhr) {
+            alert(xhr.responseText);
+          },
+          error: function(xhr, status, error) {
+            alert(xhr.responseText);
+          },
+        });
+      },
+    });
+  }
+</script>
+
+<html>
+  <p>
+    Click the button below to request an account on the DHIS 2 Trainingland.
+    Note: the email address you use with OpenEDX will be sent to the DHIS 2
+    database to create an account.
+  </p>
+  <input type="button" onClick="requestAccount();" value="Request account" />
+  <p>
+    You will receive an invitation by email with instructions on how to complete
+    the registration in DHIS 2.
+  </p>
+</html>
