+++ package.json
@@ -6,7 +6,7 @@
     "body-parser": "~1.18.3",
     "express": "~4.16.4",
     "moment": "^2.26.0",
-    "q": "^1.4.1",
+    "q": "^1.5.1",
     "request": "^2.88.2"
   }
 }
